Move props to a separate file
